LPS-78033 Abstract out logic to new check (no logic changes)
