+++ modules/util/source-formatter/src/main/java/com/liferay/source/formatter/checkstyle/checks/MissingEmptyLineCheck.java
@@ -15,11 +15,9 @@
 package com.liferay.source.formatter.checkstyle.checks;
 
 import com.liferay.portal.kernel.util.StringUtil;
-import com.liferay.source.formatter.checks.util.SourceUtil;
 import com.liferay.source.formatter.checkstyle.util.DetailASTUtil;
 
 import com.puppycrawl.tools.checkstyle.api.DetailAST;
-import com.puppycrawl.tools.checkstyle.api.FileContents;
 import com.puppycrawl.tools.checkstyle.api.TokenTypes;
 
 import java.util.List;
@@ -32,32 +30,11 @@ public class MissingEmptyLineCheck extends BaseCheck {
 
 	@Override
 	public int[] getDefaultTokens() {
-		return new int[] {TokenTypes.ASSIGN, TokenTypes.CTOR_DEF};
+		return new int[] {TokenTypes.ASSIGN};
 	}
 
 	@Override
 	protected void doVisitToken(DetailAST detailAST) {
-		if (detailAST.getType() == TokenTypes.CTOR_DEF) {
-			FileContents fileContents = getFileContents();
-
-			String fileName = StringUtil.replace(
-				fileContents.getFileName(), '\\', '/');
-
-			String absolutePath = SourceUtil.getAbsolutePath(fileName);
-
-			if (absolutePath.contains(
-					"modules/apps/forms-and-workflow/dynamic-data-mapping" +
-						"/dynamic-data-mapping-api/")) {
-
-				_checkConstructorToken(detailAST);
-			}
-		}
-		else {
-			_checkAssignToken(detailAST);
-		}
-	}
-
-	private void _checkAssignToken(DetailAST detailAST) {
 		DetailAST firstChildAST = detailAST.getFirstChild();
 
 		if ((firstChildAST == null) ||
@@ -85,57 +62,6 @@ public class MissingEmptyLineCheck extends BaseCheck {
 			parentAST, nameAST.getText(), DetailASTUtil.getEndLine(detailAST));
 		_checkMissingEmptyLineBetweenAssigningAndUsingVariable(
 			parentAST, nameAST.getText(), DetailASTUtil.getEndLine(detailAST));
-	}
-
-	private void _checkConstructorToken(DetailAST detailAST) {
-		DetailAST statementsAST = detailAST.findFirstToken(TokenTypes.SLIST);
-
-		if (statementsAST == null) {
-			return;
-		}
-
-		List<String> parameterNames = DetailASTUtil.getParameterNames(
-			detailAST);
-
-		if (parameterNames.isEmpty()) {
-			return;
-		}
-
-		DetailAST nextExpressionAST = statementsAST.getFirstChild();
-
-		if (!_isExpressionAssignsParameter(nextExpressionAST, parameterNames)) {
-			return;
-		}
-
-		int endLine = DetailASTUtil.getEndLine(nextExpressionAST);
-
-		while (true) {
-			nextExpressionAST = nextExpressionAST.getNextSibling();
-
-			nextExpressionAST = nextExpressionAST.getNextSibling();
-
-			if ((nextExpressionAST != null) &&
-				(nextExpressionAST.getType() == TokenTypes.RCURLY)) {
-
-				return;
-			}
-
-			if (!_isExpressionAssignsParameter(
-					nextExpressionAST, parameterNames)) {
-
-				int startLine = DetailASTUtil.getStartLine(nextExpressionAST);
-
-				if ((endLine + 1) != startLine) {
-					return;
-				}
-
-				log(startLine, _MSG_MISSING_EMPTY_LINE_CONSTRUCTOR, startLine);
-
-				return;
-			}
-
-			endLine = DetailASTUtil.getEndLine(nextExpressionAST);
-		}
 	}
 
 	private void _checkMissingEmptyLineAfterReferencingVariable(
@@ -274,47 +200,6 @@ public class MissingEmptyLineCheck extends BaseCheck {
 		}
 
 		return false;
-	}
-
-	private boolean _isExpressionAssignsParameter(
-		DetailAST expressionAST, List<String> parameters) {
-
-		if ((expressionAST == null) ||
-			(expressionAST.getType() != TokenTypes.EXPR)) {
-
-			return false;
-		}
-
-		DetailAST childAST = expressionAST.getFirstChild();
-
-		if (childAST.getType() != TokenTypes.ASSIGN) {
-			return false;
-		}
-
-		if (childAST.getChildCount() != 2) {
-			return false;
-		}
-
-		DetailAST firstChildAST = childAST.getFirstChild();
-		DetailAST lastChildAST = childAST.getLastChild();
-
-		if ((firstChildAST.getType() != TokenTypes.IDENT) ||
-			(lastChildAST.getType() != TokenTypes.IDENT)) {
-
-			return false;
-		}
-
-		if (!parameters.contains(lastChildAST.getText())) {
-			return false;
-		}
-
-		DetailAST nextSiblingAST = expressionAST.getNextSibling();
-
-		if (nextSiblingAST.getType() != TokenTypes.SEMI) {
-			return false;
-		}
-
-		return true;
 	}
 
 	private boolean _isExpressionAssignsVariable(
@@ -481,8 +366,5 @@ public class MissingEmptyLineCheck extends BaseCheck {
 
 	private static final String _MSG_MISSING_EMPTY_LINE_BEFORE_VARIABLE_USE =
 		"empty.line.missing.before.variable.use";
-
-	private static final String _MSG_MISSING_EMPTY_LINE_CONSTRUCTOR =
-		"empty.line.missing.constructor";
 
 }
