+++ modules/util/source-formatter/src/main/resources/checkstyle.xml
@@ -116,6 +116,9 @@
 			<message key="name.invalidConstantPattern" value="{0} constant ''{1}'' must match pattern ''{2}''" />
 			<message key="name.invalidConstantTypePattern" value="{0} constant ''{1}'' of type ''{2}'' must match pattern ''{3}''" />
 		</module>
+		<module name="com.liferay.source.formatter.checkstyle.checks.ConstructorMissingEmptyLine">
+			<message key="empty.line.missing" value="There should be an empty line before line ''{0}''" />
+		</module>
 		<module name="com.liferay.source.formatter.checkstyle.checks.DeprecatedMethodCalls">
 			<property name="enabled" value="false" />
 			<message key="method.call.deprecated" value="Method ''{0}'' has been deprecated" />
@@ -178,7 +181,6 @@
 		<module name="com.liferay.source.formatter.checkstyle.checks.MissingEmptyLine">
 			<message key="empty.line.missing.after.variable.reference" value="There should be an empty line before line ''{0}'', as we finished referencing variable ''{1}''" />
 			<message key="empty.line.missing.before.variable.use" value="There should be an empty line between assigning and using variable ''{0}''" />
-			<message key="empty.line.missing.constructor" value="There should be an empty line before line ''{0}''" />
 		</module>
 		<module name="com.liferay.source.formatter.checkstyle.checks.MissingModifier">
 			<message key="modifier.missing" value="Missing modifier for ''{0}''" />
